added confirmation dialog

+++ src/app/components/todo-list/todo-list.component.ts
@@ -1,6 +1,9 @@
 import {Component, OnInit} from '@angular/core';
 import {TodoItem} from '../../interfaces';
 import {DataProviderService} from '../../services/data-provider.service';
+import {MatDialog} from '@angular/material';
+import {ConfirmationDialogComponent} from '../dialogs/confirmation-dialog/confirmation-dialog.component';
+import {DialogTypes} from '../../enums';
 
 @Component({
   selector: 'app-todo-list',
@@ -11,7 +14,7 @@ export class TodoListComponent implements OnInit {
   dataSource: TodoItem[] = [];
   columnsToShow: string [] = ['position', 'name', 'createdAt', 'editedAt', 'action'];
 
-  constructor(private dataProviderService: DataProviderService) {
+  constructor(private dataProviderService: DataProviderService, private dialog: MatDialog) {
   }
 
   ngOnInit(): void {
@@ -26,12 +29,22 @@ export class TodoListComponent implements OnInit {
 
   deleteRecord(event, tableRow: TodoItem): void {
     event.stopPropagation();
-    this.dataProviderService.deleteItem(tableRow.id)
-      .then(result => {
-        const index = this.dataSource.findIndex(el => el.id === tableRow.id);
-        this.dataSource.splice(index, 1);
-        this.dataSource.slice();
-      });
+
+    const dialog = this.dialog.open(ConfirmationDialogComponent, {
+      hasBackdrop: false,
+      data: {type: DialogTypes.DELETE}
+    });
+
+    dialog.afterClosed().subscribe(isConfirmed => {
+      if (isConfirmed) {
+        this.dataProviderService.deleteItem(tableRow.id)
+          .then(result => {
+            const index = this.dataSource.findIndex(el => el.id === tableRow.id);
+            this.dataSource.splice(index, 1);
+            this.dataSource = this.dataSource.slice();
+          });
+      }
+    });
   }
 
   private updateData() {
